Created CreateJabberPhone.py
This is the python script automationtool.hta uses to create the phone. It is called from the script with command line arguments and sends AXL requests to the specified call manager. Communcation after it's called is one way to the script and uses windows pop up boxes. It also writes errors to afile named CreatePhoneErrorFile.txt in it's local directory.

The actual call manager addresses have been as they can be considered internal company information.

This is my first major programing experiance outside of school as well as my first major Python project. It's over a year old now and I can still understand it easily enough which I take as a positive. Some of the comments are a bit whimsical, and I wrote a large part of it without writing any functions which made it longer than it needed to be. The line spacing was also ruined when I send it over here.