Create CreateJabberPhone.py
This is the python script AutomationTool.hta uses to create the phone. It is called from the application with command line arguments and sends AXL requests to the specified call manager. Communication after it is called is one way to the script and uses Windows pop up boxes. It also writes errors to a file named CreatePhoneErrorFile.txt in it's local directory.

The actual call manager addresses have been removed as they can be considered internal company information.

This is my first major programming experience outside of school as well as my first major Python project. It has been over a year since I made this and I can still understand it easily enough, which I take as a positive. Some of the comments are a bit whimsical, and I wrote a large part of it without writing any functions which made it longer than it needed to be. The line spacing was also ruined when I sent it over here.